[rename/#66] 드롭다운 컴포넌트 이름 수정
